refactor: Modulariser session_detail_screen.dart (556L → 204L + 477L components)
- Extraction: SessionHeaderCard, SessionCoachAnalysisSection, SessionExercisesSection, SessionSyntheseSection
- Extraction widgets: SessionChip, StatBlock, DividerVert, LoggedCoachAnalysis
- Fichier: lib/screens/session_detail/session_detail_components.dart (477L)
- Suppression classes internes obsolètes
- Compilation validée (mcp_dart_sdk_mcp__analyze_files)
- Référence: Phase 5 de v0.4 - Tâche T3 refactoring
