+++ docs/specs/v0.4/v0.4_scope.md
@@ -4,19 +4,19 @@ Release technique, avec outillage, refactoring, changement d'architecture, authe
 
 ## v0.4 : Statut des développements
 
-| ID  | Nom feature / tâche principale                            | Lot         | Priorité  | Dépendances  | Statut      |
-|------|----------------------------------------------------------|-------------|-----------|--------------|-------------|
-| T1  | Intégration SonarCloud                                    | Technique   | Haute     | -            | **Fait**    |
-| T2  | Cahier de recette (tests manuels)                         | Technique   | Moyenne   | T1           | **Fait**    |
-| T3  | Refactoring sans régression                               | Technique   | Haute     | T1, T2       | En cours    |
-| T4  | Authentification via NexTarget-server                     | Technique   | Moyenne   | T3           | A faire     |
-| T5  | Déport des appels Mistral API vers NexTarget-server       | Technique   | Moyenne   | T3           | À faire     |
-| D1  | Dashboard harmonisation et graphiques multi-courbes       | Dashboard   | Haute     | -            | **Fait**    |
-| D2  | Graphiques évolution par méthode de prise (1/2 mains)     | Dashboard   | Haute     | D1           | **Fait**    |
-| D3  | Statistique "Prise dominante" (remplacement catégories)   | Dashboard   | Moyenne   | D1           | **Fait**    |
-| F1  | Règles de sécurité FFTir (révision contenu & affichage)   | Feature     | Basse     | -            | À faire     |
-| F2  | Duplication d'un Exercice                                 | Feature     | Basse     | -            | À faire     |
-| F3  | Graphe d'une Session réalisée                             | Feature     | Basse     | D1           | À faire     |
+| ID  | Nom feature / tâche principale                            | Lot         | Priorité  | Dépendances  | Statut         |
+|------|----------------------------------------------------------|-------------|-----------|--------------|----------------|
+| T1  | Intégration SonarCloud                                    | Technique   | Haute     | -            | **✅ FAIT**    |
+| T2  | Cahier de recette (tests manuels)                         | Technique   | Moyenne   | T1           | **✅ FAIT**    |
+| T3  | Refactoring sans régression                               | Technique   | Haute     | T1, T2       | **🟡 À FAIRE** |
+| T4  | Authentification via NexTarget-server                     | Technique   | Moyenne   | T3           | A faire        |
+| T5  | Déport des appels Mistral API vers NexTarget-server       | Technique   | Moyenne   | T3           | À faire        |
+| D1  | Dashboard harmonisation et graphiques multi-courbes       | Dashboard   | Haute     | -            | **✅ FAIT**    |
+| D2  | Graphiques évolution par méthode de prise (1/2 mains)     | Dashboard   | Haute     | D1           | **✅ FAIT**    |
+| D3  | Statistique "Prise dominante" (remplacement catégories)   | Dashboard   | Moyenne   | D1           | **✅ FAIT**    |
+| F1  | Règles de sécurité FFTir (révision contenu & affichage)   | Feature     | Basse     | -            | À faire        |
+| F2  | Duplication d'un Exercice                                 | Feature     | Basse     | -            | À faire        |
+| F3  | Graphe d'une Session réalisée                             | Feature     | Basse     | D1           | À faire        |
 
 ## Ajouts Techniques
